requirements.txt
python-decouple
ipython
selenium
bs4
pandas